backend/requirements.txt
# FastAPI Core
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.4
pydantic-settings==2.6.1

# Database
sqlmodel==0.0.22
asyncpg==0.30.0
psycopg2-binary==2.9.10
alembic==1.14.0
greenlet==3.1.1
python-dotenv==1.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1